Fix Python2/3 import logic, and cleanup
